Add more security measures: 1. check if user owns a post or comment 2. check if blog or comment id is valid

+++ multi-user-blog/python/webpage_backend/blog_page/blog_page_deletecomment.py
@@ -3,7 +3,9 @@ from python.webpage_backend.webpage_handler import *
 class BlogDeleteCommentHandler(Handler):
 
 	def post(self):
+		super(BlogDeleteCommentHandler, self).check_valid_user(self.logged_in_user)
 		comment_id = int(self.request.get('comment-id'))
 		super(BlogDeleteCommentHandler, self).check_valid_comment_id(comment_id)
-		Comment.get_by_id(comment_id, parent = self.root_comment_key).delete()
+		if super(BlogDeleteCommentHandler, self).check_user_owns_comment(comment_id):
+			Comment.get_by_id(comment_id, parent = self.root_comment_key).delete()
 		self.redirect('/blog')
